+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/timeline/Timeline.java
@@ -66,7 +66,7 @@ public class Timeline {
 
     public RECT displayRect;
 
-    public int frameRate;
+    public float frameRate;
 
     public Timelined timelined;
